Additional Features
Sending a direct message to a bad client ID does not cause a warning (10 points).
Sending a direct message to a bad client ID notifies the sender that the message could not be sent (5 points).
Change the greeting message from the host so it is not a whisper from the host, but just a message sent directly to the connected client (5 points).
Send a message to all clients when a client connects (5 points).
Send a message to all clients when a client disconnects (5 points).